Create Lobby , also fix some bugs with signals

+++ main_home/signals.py
@@ -3,7 +3,7 @@ from django.dispatch import receiver
 from django.core.files.storage import default_storage
 import os
 
-from main_home.models import Appointment
+from main_home.models import Appointment, Lobby
 
 from django.db.models.signals import pre_save, pre_delete
 from client.models import Client
@@ -84,12 +84,13 @@ def delete_unused_profile_pics():
 @receiver(pre_save, sender=Receptionist)
 @receiver(pre_save, sender=Auditor)
 def handle_user_profile_pic_change(sender, instance, **kwargs):
-    if instance.pk:
+    if sender.objects.filter(pk=instance.pk).exists() :
         # Retrieve the existing profile picture path
         old_profile_pic_path = sender.objects.get(pk=instance.pk).profile_pic.path
+        old_profile_pic = sender.objects.get(pk=instance.pk).profile_pic
 
         # Check if the existing profile picture is different from the new profile picture
-        if old_profile_pic_path != instance.profile_pic.path:
+        if old_profile_pic_path != instance.profile_pic.path and old_profile_pic != "default.png" :
             # Delete the old profile picture
             default_storage.delete(old_profile_pic_path)
 
@@ -101,8 +102,22 @@ def handle_user_profile_pic_change(sender, instance, **kwargs):
 def handle_user_profile_pic_deletion(sender, instance, **kwargs):
     # Delete the profile picture when a user is deleted
     profile_pic_path = instance.profile_pic.path
-    default_storage.delete(profile_pic_path)
+    profile_pic = instance.profile_pic
+    if profile_pic != "default.png" :
+        default_storage.delete(profile_pic_path)
 
     # Delete unused profile pictures
     delete_unused_profile_pics()
 
+
+# Nurse Lobby
+@receiver(post_save, sender=Nurse)
+def create_nurse_lobby(sender, instance, created, **kwargs):
+    if created:
+        lobby = Lobby.objects.create(nurse=instance)
+        print(lobby.clients)
+        
+@receiver(pre_delete, sender=Nurse)
+def delete_nurse_lobby(sender, instance, **kwargs):
+    if instance.lobby:
+        instance.lobby.delete()
